requirements.txt
loguru==0.7.2
matplotlib==3.8.0
metabase_api==0.3.1
numpy==1.26.0
pandas==2.1.1
psycopg2_binary==2.9.7
python-dotenv==1.0.0
seaborn==0.12.2
SQLAlchemy==2.0.21